requirements.txt
streamlit
streamlit-webrtc
opencv-python-headless
mediapipe
numpy
pillow
av